+++ src/app/(auth)/login/page.tsx
@@ -5,8 +5,8 @@ import { useRouter } from 'next/navigation';
 import Link from 'next/link';
 import { useMutation } from '@tanstack/react-query';
 import { useAuth } from '@/app/services/auth-provider'; // Zustand 대신 새 Provider 사용
-import SuccessModal from '@/app/components/SuccessModal';
-import FailModal from '@/app/components/FailModal';
+import SuccessModal from '@/app/_components/SuccessModal';
+import FailModal from '@/app/_components/FailModal';
 import { PATHS } from '@/app/(auth)/login/types/auth';
 
 // JWT 로그인 API 함수
