+++ src/app/_components/Navbar.tsx
@@ -4,7 +4,7 @@ import Link from 'next/link';
 import { useState, useRef, useEffect } from 'react';
 import NotificationList from '@/app/components/NotificationList';
 import { useAuth } from '@/app/services/auth-provider'; // 새 인증 Provider 사용
-import type { Notification } from '@/app/types/notification.types';
+import type { Notification } from '@/app/_types/notification.types';
 
 export default function TopNavigation() {
   const [showNotifications, setShowNotifications] = useState(false);
