+++ src/app/_types/notification.types.ts
